+++ scripts/solve_network.py
@@ -159,6 +159,9 @@ def _add_land_use_constraint_m(n, planning_horizons, config):
     current_horizon = snakemake.wildcards.planning_horizons
 
     for carrier in ["solar", "onwind", "offwind-ac", "offwind-dc"]:
+        extendable_i = (n.generators.carrier == carrier) & n.generators.p_nom_extendable
+        n.generators.loc[extendable_i, "p_nom_min"] = 0
+
         existing = n.generators.loc[n.generators.carrier == carrier, "p_nom"]
         ind = list(
             {i.split(sep=" ")[0] + " " + i.split(sep=" ")[1] for i in existing.index}
@@ -179,6 +182,19 @@ def _add_land_use_constraint_m(n, planning_horizons, config):
             n.generators.loc[sel_current, "p_nom_max"] -= existing.loc[
                 sel_p_year
             ].rename(lambda x: x[:-4] + current_horizon)
+
+    # check if existing capacities are larger than technical potential
+    existing_large = n.generators[
+        n.generators["p_nom_min"] > n.generators["p_nom_max"]
+    ].index
+    if len(existing_large):
+        logger.warning(
+            f"Existing capacities larger than technical potential for {existing_large},\
+                        adjust technical potential to existing capacities"
+        )
+        n.generators.loc[existing_large, "p_nom_max"] = n.generators.loc[
+            existing_large, "p_nom_min"
+        ]
 
     n.generators.p_nom_max.clip(lower=0, inplace=True)